Improve visual appearance
Now it doesn't look incredibly ugly. This was achieved by manually
drawing keys rather than using the built-in drawing functionality.

I also got rid of the popup-style preview and replaced it with a feature
that immediately displays the non-long-press version of a key within the
input field when the key is pressed, and changes it to the long-press
version if the key is held.